+++ src/routes/(components)/select.svelte
@@ -27,7 +27,7 @@
 	class="flex w-[140px] items-center justify-between rounded-md border border-magnum-700 bg-neutral-800 px-3 py-1 text-magnum-600
   outline-none transition-opacity hover:opacity-75 focus:!border-magnum-400 focus:!text-magnum-400"
 	{...$trigger}
-	use:trigger.action
+	use:trigger
 	aria-label="Select"
 >
 	{$label || 'Select an option'}
@@ -38,14 +38,14 @@
 	class="0 z-10 flex max-h-[360px] flex-col
   overflow-y-auto rounded-md bg-neutral-800 p-1 shadow-md focus:!ring-0"
 	{...$menu}
-	use:menu.action
+	use:menu
 >
 	{#each options as o}
 		<li
 			class="relative cursor-pointer rounded-md py-1 pl-8 pr-4 text-neutral-500 outline-none
 			focus:!text-magnum-400 data-[disabled]:pointer-events-none data-[disabled]:cursor-not-allowed data-[selected]:text-magnum-600 data-[disabled]:line-through"
 			{...$option({ ...o })}
-			use:option.action
+			use:option
 		>
 			{#if $isSelected(o.value)}
 				<div class="check">