+++ src/routes/old/builders/dialog/(examples)/(drawer)/index.ts
@@ -1,7 +1,8 @@
+import Preview from './tailwind.svelte';
 import type { PreviewProps } from '$routes/(components)';
+
 import tw from './tailwind.svelte?raw';
 import TwConfig from '$routes/(configs)/tailwind.config.ts?raw';
-import Preview from './tailwind.svelte';
 
 const Tailwind: PreviewProps['code']['Tailwind'] = {
 	'index.svelte': tw,