+++ src/routes/old/builders/dialog/(examples)/(preview)/index.ts
@@ -1,6 +1,6 @@
+import Preview from './tailwind.svelte';
 import type { PreviewProps } from '$routes/(components)';
 
-import Preview from './tailwind.svelte';
 import tw from './tailwind.svelte?raw';
 import TwConfig from '$routes/(configs)/tailwind.config.ts?raw';